Fix #19227: Preferences changes
