+++ src/appshell/CMakeLists.txt
@@ -69,14 +69,14 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/folderspreferencesmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/canvaspreferencesmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/canvaspreferencesmodel.h
+    ${CMAKE_CURRENT_LIST_DIR}/view/preferences/saveandpublishpreferencesmodel.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/view/preferences/saveandpublishpreferencesmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/generalpreferencesmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/generalpreferencesmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/updatepreferencesmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/updatepreferencesmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/appearancepreferencesmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/appearancepreferencesmodel.h
-    ${CMAKE_CURRENT_LIST_DIR}/view/preferences/programmestartpreferencesmodel.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/view/preferences/programmestartpreferencesmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/noteinputpreferencesmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/noteinputpreferencesmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/preferences/advancedpreferencesmodel.cpp
