+++ src/appshell/appshell.qrc
@@ -28,7 +28,6 @@
         <file>qml/Preferences/PreferencesPage.qml</file>
         <file>qml/Preferences/GeneralPreferencesPage.qml</file>
         <file>qml/Preferences/UpdatePreferencesPage.qml</file>
-        <file>qml/Preferences/ProgrammeStartPreferencesPage.qml</file>
         <file>qml/Preferences/NoteInputPreferencesPage.qml</file>
         <file>qml/Preferences/AppearancePreferencesPage.qml</file>
         <file>qml/Preferences/FoldersPreferencesPage.qml</file>
@@ -39,7 +38,7 @@
         <file>qml/Preferences/ImportPreferencesPage.qml</file>
         <file>qml/Preferences/MidiDeviceMappingPreferencesPage.qml</file>
         <file>qml/Preferences/IOPreferencesPage.qml</file>
-        <file>qml/Preferences/CloudPreferencesPage.qml</file>
+        <file>qml/Preferences/SaveAndPublishPreferencesPage.qml</file>
         <file>qml/Preferences/PreferencesDialog.qml</file>
         <file>qml/Preferences/internal/PreferencesButtonsPanel.qml</file>
         <file>qml/Preferences/internal/PreferencesMenu.qml</file>
@@ -59,7 +58,7 @@
         <file>qml/Preferences/internal/ComboBoxWithTitle.qml</file>
         <file>qml/Preferences/internal/IncrementalPropertyControlWithTitle.qml</file>
         <file>qml/Preferences/internal/KeyboardLayoutsSection.qml</file>
-        <file>qml/Preferences/internal/AudioGenerationSection.qml</file>
+        <file>qml/Preferences/internal/SaveToCloudSection.qml</file>
         <file>qml/DevTools/KeyNav/KeyNavExample.qml</file>
         <file>qml/DevTools/KeyNav/KeyNavSection.qml</file>
         <file>qml/DevTools/KeyNav/KeyNavSubSection.qml</file>
@@ -88,7 +87,7 @@
         <file>qml/Preferences/internal/LanguagesSection.qml</file>
         <file>qml/Preferences/internal/AutoSaveSection.qml</file>
         <file>qml/Preferences/internal/RemoteControlSection.qml</file>
-        <file>qml/Preferences/internal/ProgrammeStartSection.qml</file>
+        <file>qml/Preferences/internal/ProgramStartSection.qml</file>
         <file>qml/Preferences/internal/FoldersSection.qml</file>
         <file>qml/Preferences/internal/ThemeAdditionalOptionsSection.qml</file>
         <file>qml/Preferences/internal/AdvancedTopSection.qml</file>
@@ -110,5 +109,6 @@
         <file>qml/Preferences/internal/BrailleSection.qml</file>
         <file>qml/Preferences/internal/BrailleTableSection.qml</file>
         <file>qml/Preferences/internal/MeiSection.qml</file>
+        <file>qml/Preferences/internal/PublishMuseScoreComSection.qml</file>
     </qresource>
 </RCC>
