+++ src/appshell/qml/Preferences/GeneralPreferencesPage.qml
@@ -67,6 +67,27 @@ PreferencesPage {
 
         SeparatorLine { }
 
+        ProgramStartSection {
+            startupModes: preferencesModel.startupModes
+            scorePathFilter: preferencesModel.scorePathFilter()
+            panels: preferencesModel.panels
+
+            navigation.section: root.navigationSection
+            navigation.order: root.navigationOrderStart + 2
+
+            onCurrentStartupModesChanged: function(index) {
+                preferencesModel.setCurrentStartupMode(index)
+            }
+
+            onStartupScorePathChanged: function(path) {
+                preferencesModel.setStartupScorePath(path)
+            }
+
+            onPanelsVisibleChanged: function(panelIndex, visible) {
+                preferencesModel.setPanelVisible(panelIndex, visible)
+            }
+        }
+
         /*
          * TODO: https://github.com/musescore/MuseScore/issues/9807
         KeyboardLayoutsSection {
@@ -83,22 +104,6 @@ PreferencesPage {
 
         SeparatorLine { }
         */
-
-        AutoSaveSection {
-            isAutoSaveEnabled: preferencesModel.isAutoSaveEnabled
-            autoSaveInterval: preferencesModel.autoSaveInterval
-
-            navigation.section: root.navigationSection
-            navigation.order: root.navigationOrderStart + 3
-
-            onAutoSaveEnabledChanged: function(enabled) {
-                preferencesModel.isAutoSaveEnabled = enabled
-            }
-
-            onIntervalChanged: function(minutes) {
-                preferencesModel.autoSaveInterval = minutes
-            }
-        }
 
         /*
          * TODO: https://github.com/musescore/MuseScore/issues/9807
