+++ src/appshell/qml/Preferences/internal/ProgramStartSection.qml
