+++ src/appshell/qml/Preferences/internal/PublishMuseScoreComSection.qml
@@ -5,7 +5,7 @@
  * MuseScore
  * Music Composition & Notation
  *
- * Copyright (C) 2021 MuseScore BVBA and others
+ * Copyright (C) 2023 MuseScore BVBA and others
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as
@@ -21,13 +21,30 @@
  */
 import QtQuick 2.15
 
-import "internal"
+import MuseScore.Ui 1.0
+import MuseScore.UiComponents 1.0
 
-PreferencesPage {
+BaseSection {
     id: root
 
-    AudioGenerationSection {
-        navigation.section: root.navigationSection
-        navigation.order: root.navigationOrderStart + 1
+    title: qsTrc("appshell/preferences", "Publish to MuseScore.com")
+
+    property alias promptShareAudioCom: promptShareAudioComCheckBox.checked
+
+    signal promptShareAudioComChangeRequested(bool prompt)
+
+    CheckBox {
+        id: promptShareAudioComCheckBox
+
+        width: parent.width
+
+        text: qsTrc("appshell/preferences", "Always prompt to share on Audio.com after publishing to MuseScore.com")
+
+        navigation.name: "PromptShareAudioComCheckBox"
+        navigation.panel: root.navigation
+
+        onClicked: {
+            root.promptShareAudioComChangeRequested(!checked)
+        }
     }
 }
