+++ src/appshell/view/preferences/generalpreferencesmodel.h
@@ -27,6 +27,7 @@
 #include "progress.h"
 
 #include "modularity/ioc.h"
+#include "iappshellconfiguration.h"
 #include "async/asyncable.h"
 
 #include "global/iinteractive.h"
@@ -40,11 +41,11 @@ class GeneralPreferencesModel : public QObject, public async::Asyncable
 {
     Q_OBJECT
 
+    INJECT(IAppShellConfiguration, configuration)
     INJECT(framework::IInteractive, interactive)
     INJECT(languages::ILanguagesConfiguration, languagesConfiguration)
     INJECT(languages::ILanguagesService, languagesService)
     INJECT(shortcuts::IShortcutsConfiguration, shortcutsConfiguration)
-    INJECT(project::IProjectConfiguration, projectConfiguration)
 
     Q_PROPERTY(QVariantList languages READ languages NOTIFY languagesChanged)
     Q_PROPERTY(QString currentLanguageCode READ currentLanguageCode WRITE setCurrentLanguageCode NOTIFY currentLanguageCodeChanged)
@@ -52,12 +53,13 @@ class GeneralPreferencesModel : public QObject, public async::Asyncable
     Q_PROPERTY(QStringList keyboardLayouts READ keyboardLayouts CONSTANT)
     Q_PROPERTY(QString currentKeyboardLayout READ currentKeyboardLayout WRITE setCurrentKeyboardLayout NOTIFY currentKeyboardLayoutChanged)
 
-    Q_PROPERTY(bool isAutoSaveEnabled READ isAutoSaveEnabled WRITE setAutoSaveEnabled NOTIFY autoSaveEnabledChanged)
-    Q_PROPERTY(int autoSaveInterval READ autoSaveInterval WRITE setAutoSaveInterval NOTIFY autoSaveIntervalChanged)
     Q_PROPERTY(bool isOSCRemoteControl READ isOSCRemoteControl WRITE setIsOSCRemoteControl NOTIFY isOSCRemoteControlChanged)
     Q_PROPERTY(int oscPort READ oscPort WRITE setOscPort NOTIFY oscPortChanged)
 
     Q_PROPERTY(bool isNeedRestart READ isNeedRestart WRITE setIsNeedRestart NOTIFY isNeedRestartChanged)
+
+    Q_PROPERTY(QVariantList startupModes READ startupModes NOTIFY startupModesChanged)
+    Q_PROPERTY(QVariantList panels READ panels NOTIFY panelsChanged)
 
 public:
     explicit GeneralPreferencesModel(QObject* parent = nullptr);
@@ -65,14 +67,21 @@ public:
     Q_INVOKABLE void load();
     Q_INVOKABLE void checkUpdateForCurrentLanguage();
 
+    Q_INVOKABLE void setCurrentStartupMode(int modeIndex);
+    Q_INVOKABLE void setStartupScorePath(const QString& scorePath);
+    Q_INVOKABLE void setPanelVisible(int panelIndex, bool visible);
+
+    Q_INVOKABLE QStringList scorePathFilter() const;
+
     QVariantList languages() const;
     QString currentLanguageCode() const;
 
     QStringList keyboardLayouts() const;
     QString currentKeyboardLayout() const;
 
-    bool isAutoSaveEnabled() const;
-    int autoSaveInterval() const;
+    QVariantList startupModes() const;
+    QVariantList panels() const;
+
     bool isOSCRemoteControl() const;
     int oscPort() const;
     bool isNeedRestart() const;
@@ -80,8 +89,6 @@ public:
 public slots:
     void setCurrentLanguageCode(const QString& currentLanguageCode);
     void setCurrentKeyboardLayout(const QString& keyboardLayout);
-    void setAutoSaveEnabled(bool enabled);
-    void setAutoSaveInterval(int minutes);
     void setIsOSCRemoteControl(bool isOSCRemoteControl);
     void setOscPort(int oscPort);
     void setIsNeedRestart(bool newIsNeedRestart);
@@ -90,18 +97,48 @@ signals:
     void languagesChanged(QVariantList languages);
     void currentLanguageCodeChanged(QString currentLanguageCode);
     void currentKeyboardLayoutChanged();
-    void autoSaveEnabledChanged(bool enabled);
-    void autoSaveIntervalChanged(int minutes);
     void isOSCRemoteControlChanged(bool isOSCRemoteControl);
     void oscPortChanged(int oscPort);
 
     void receivingUpdateForCurrentLanguage(int current, int total, QString status);
     void isNeedRestartChanged();
 
+    void startupModesChanged();
+    void panelsChanged();
+
 private:
     framework::Progress m_languageUpdateProgress;
 
     bool m_isNeedRestart = false;
+
+    enum PanelType {
+        Unknown,
+        SplashScreen,
+        Navigator
+    };
+
+    struct Panel
+    {
+        PanelType type = Unknown;
+        QString title;
+        bool visible = false;
+    };
+
+    using PanelList = QList<Panel>;
+
+    struct StartMode
+    {
+        StartupModeType type = StartupModeType::StartWithNewScore;
+        QString title;
+        bool checked = false;
+        bool canSelectScorePath = false;
+        QString scorePath;
+    };
+
+    using StartModeList = QList<StartMode>;
+
+    PanelList allPanels() const;
+    StartModeList allStartupModes() const;
 };
 }
