+++ src/appshell/view/preferences/preferencesmodel.cpp
@@ -147,17 +147,9 @@ void PreferencesModel::load(const QString& currentPageId)
 
     m_rootItem = new PreferencePageItem();
 
-    QList<PreferencePageItem*> generalItems {
-        makeItem("general-start", QT_TRANSLATE_NOOP("appshell/preferences", "Program start"), IconCode::Code::NONE,
-                 "Preferences/ProgrammeStartPreferencesPage.qml"),
-
-        makeItem("general-folders", QT_TRANSLATE_NOOP("appshell/preferences", "Folders"), IconCode::Code::NONE,
-                 "Preferences/FoldersPreferencesPage.qml")
-    };
-
     QList<PreferencePageItem*> items {
         makeItem("general", QT_TRANSLATE_NOOP("appshell/preferences", "General"), IconCode::Code::SETTINGS_COG,
-                 "Preferences/GeneralPreferencesPage.qml", generalItems),
+                 "Preferences/GeneralPreferencesPage.qml"),
 
         makeItem("appearance", QT_TRANSLATE_NOOP("appshell/preferences", "Appearance"), IconCode::Code::EYE_OPEN,
                  "Preferences/AppearancePreferencesPage.qml"),
@@ -165,8 +157,8 @@ void PreferencesModel::load(const QString& currentPageId)
         makeItem("canvas", QT_TRANSLATE_NOOP("appshell/preferences", "Canvas"), IconCode::Code::NEW_FILE,
                  "Preferences/CanvasPreferencesPage.qml"),
 
-        makeItem("cloud", QT_TRANSLATE_NOOP("appshell/preferences", "Cloud"), IconCode::Code::CLOUD_FILE,
-                 "Preferences/CloudPreferencesPage.qml"),
+        makeItem("cloud", QT_TRANSLATE_NOOP("appshell/preferences", "Save & publish"), IconCode::Code::CLOUD_FILE,
+                 "Preferences/SaveAndPublishPreferencesPage.qml"),
 
         makeItem("note-input", QT_TRANSLATE_NOOP("appshell/preferences", "Note input"), IconCode::Code::EDIT,
                  "Preferences/NoteInputPreferencesPage.qml"),
@@ -188,6 +180,9 @@ void PreferencesModel::load(const QString& currentPageId)
 
         makeItem("update", QT_TRANSLATE_NOOP("appshell/preferences", "Update"), IconCode::Code::UPDATE,
                  "Preferences/UpdatePreferencesPage.qml"),
+
+        makeItem("general-folders", QT_TRANSLATE_NOOP("appshell/preferences", "Folders"), IconCode::Code::OPEN_FILE,
+                 "Preferences/FoldersPreferencesPage.qml"),
 
         makeItem("advanced", QT_TRANSLATE_NOOP("appshell/preferences", "Advanced"), IconCode::Code::CONFIGURE,
                  "Preferences/AdvancedPreferencesPage.qml"),
