+++ src/project/internal/projectconfiguration.cpp
@@ -53,6 +53,7 @@ static const Settings::Key PREFERRED_SCORE_CREATION_MODE_KEY(module_name, "proje
 static const Settings::Key MIGRATION_OPTIONS(module_name, "project/migration");
 static const Settings::Key AUTOSAVE_ENABLED_KEY(module_name, "project/autoSaveEnabled");
 static const Settings::Key AUTOSAVE_INTERVAL_KEY(module_name, "project/autoSaveInterval");
+static const Settings::Key ALWAYS_PROMPT_SHARE_AUDIO_COM_AFTER_PUBLISH(module_name, "project/promptShareAudioCom");
 static const Settings::Key SHOULD_DESTINATION_FOLDER_BE_OPENED_ON_EXPORT(module_name, "project/shouldDestinationFolderBeOpenedOnExport");
 static const Settings::Key OPEN_DETAILED_PROJECT_UPLOADED_DIALOG(module_name, "project/openDetailedProjectUploadedDialog");
 static const Settings::Key HAS_ASKED_AUDIO_GENERATION_SETTINGS(module_name, "project/hasAskedAudioGenerationSettings");
@@ -95,6 +96,11 @@ void ProjectConfiguration::init()
     settings()->setDefaultValue(AUTOSAVE_INTERVAL_KEY, Val(5));
     settings()->valueChanged(AUTOSAVE_INTERVAL_KEY).onReceive(nullptr, [this](const Val& val) {
         m_autoSaveIntervalChanged.send(val.toInt());
+    });
+
+    settings()->setDefaultValue(ALWAYS_PROMPT_SHARE_AUDIO_COM_AFTER_PUBLISH, Val(false));
+    settings()->valueChanged(ALWAYS_PROMPT_SHARE_AUDIO_COM_AFTER_PUBLISH).onReceive(nullptr, [this](const Val& val) {
+        m_promptShareAudioComChanged.send(val.toBool());
     });
 
     settings()->setDefaultValue(SHOULD_DESTINATION_FOLDER_BE_OPENED_ON_EXPORT, Val(false));
@@ -523,6 +529,21 @@ void ProjectConfiguration::setAutoSaveInterval(int minutes)
 async::Channel<int> ProjectConfiguration::autoSaveIntervalChanged() const
 {
     return m_autoSaveIntervalChanged;
+}
+
+bool ProjectConfiguration::promptShareAudioCom() const
+{
+    return settings()->value(ALWAYS_PROMPT_SHARE_AUDIO_COM_AFTER_PUBLISH).toBool();
+}
+
+void ProjectConfiguration::setPromptShareAudioCom(bool prompt)
+{
+    settings()->setSharedValue(ALWAYS_PROMPT_SHARE_AUDIO_COM_AFTER_PUBLISH, Val(prompt));
+}
+
+async::Channel<bool> ProjectConfiguration::promptShareAudioComChanged() const
+{
+    return m_promptShareAudioComChanged;
 }
 
 io::path_t ProjectConfiguration::newProjectTemporaryPath() const
